Introduce the gore pattern generator
In order to design arbitrary volume balloons the gore
pattern of a sphere must be computed. The addition of the
src/gore/gore.py program currently generates template images
that can be used for smaller balloon designs. For larger
balloon designs, src/gore/gore.py will need to be updated
to cut printer templates up, or to produce different type
of template design.